+++ src/App.js
@@ -1,14 +1,18 @@
 import React from "react";
 
-import { ContextData, MiniContextProvider } from "storage/context";
+// import { ContextData, MiniContextProvider } from "storage/context";
 import { AppLayout } from "./AppLayout";
 
+// Mode: No ContextAPI below
 function App() {
-    return (
-        <MiniContextProvider value={ContextData}>
-            <AppLayout />
-        </MiniContextProvider>
-    );
+    return <AppLayout />;
 }
+
+// Mode: Using the ContextAPI below
+// function App() {
+//     return <MiniContextProvider value={ContextData}>
+//          <AppLayout />
+//      </MiniContextProvider>;
+// }
 
 export default App;