+++ CHANGELOG.md
@@ -9,13 +9,13 @@
 -   [DONE] Created Header.jsx component with proper styling.
 -   [DONE] Created Routes.js file to centralize Route table.
 
-### Changes Planned
+### Changes Alanned
 
 -   [TODO] Create new page Profile.
 
 ### 17/11/2022 - React Session # 9
 
-#### Changes Planned
+#### Changes Applied
 
 -   [DONE][0.3h] Create a contextAPI module to be used as a hook to other components.
 -   [WIP][0.1h] Consume the ContextAPI in a component
@@ -23,12 +23,29 @@
 -   [TODO][0.3h] Querying and storing data retrieved from an API.
 -   [TODO][1.0h] Create new page Profile.
 
-### 22/11/2022 - React Review # 10
+### 22/11/2022 - React Session # 10
 
-#### Changes applied
+#### Changes Applied
 
 -   [DONE] Created AppLayout component to add another Layout for contextAPI.
 -   [DONE] Worked on modifications on HomePage to display data.
 -   [DONE] Worked on Header logic to display profile.name.
 -   [DONE] Created UserProfile file & class.
 -   [DONE] Modified storage/context/index file to add new functions.
+
+### 29/11/2022 - React Session # 11
+
+#### Changes Planned
+
+-   [WIP][0.1h] Consume the State Layer in a component via hook
+-   [WIP][0.3h] Display the data stored in the app state hook.
+-   [TODO][0.3h] Querying and storing data retrieved from an API.
+-   [TODO][1.0h] Create new page Profile.
+
+#### Changes Applied
+
+-   [DONE] Removed MiniContextProvider form App.js
+-   [DONE] Modified AppLayout in order to consume new store.
+-   [DONE] Modified Header in order to consume new store.
+-   [DONE] Created useAppState hook.
+-   [DONE] Modified StateService service to use new store.